Update tests and remove unused items for now
