+++ kafka/schema-registry-json-schema/src/main/kotlin/io/violabs/mimir/schemaregistry/jsonschema/config/AppConfig.kt
@@ -6,10 +6,5 @@ import org.springframework.context.annotation.Configuration
 @Configuration
 @EnableConfigurationProperties(
     KafkaConfigProps::class,
-//    KafkaConfigProps.Producer::class,
-//    KafkaConfigProps.Producer.Properties::class,
-//    KafkaConfigProps.Consumer::class,
-//    KafkaConfigProps.Consumer.Properties::class,
-//    KafkaConfigProps.Listener::class
 )
 class AppConfig
