+++ kafka/schema-registry-json-schema/src/main/kotlin/io/violabs/mimir/schemaregistry/jsonschema/domain/CreateUserEventRequest.kt
@@ -1,0 +1,8 @@
+package io.violabs.mimir.schemaregistry.jsonschema.domain
+
+data class CreateUserEventRequest(
+    val userId: String,
+    val eventType: EventType,
+    val userData: UserData? = null,
+    val metadata: Map<String, Any> = emptyMap()
+)
