+++ kafka/schema-registry-json-schema/src/main/resources/application.yml
@@ -23,6 +23,7 @@ spring:
       properties:
         schema.registry.url: http://localhost:8090
         json.fail.invalid.schema: true
+        json.value.type: io.violabs.mimir.schemaregistry.jsonschema.domain.UserEventV1
         specific.json.reader: true
     listener:
       ack-mode: manual
